move arm_access_support
