+++ armstrong/core/arm_access/tests/__init__.py
@@ -1,3 +1,6 @@
+from pkgutil import extend_path
+__path__ = extend_path(__path__, __name__)
+
 from .access_memberships import *
 from .fields import *
 from .subscription import *
