+++ armstrong/core/arm_access/tests/arm_access_support/models.py
@@ -1,6 +1,6 @@
 from django.db import models
-from ..models import AccessObject
-from ..mixins import AccessMixin
+from ...models import AccessObject
+from ...mixins import AccessMixin
 
 
 class Content(AccessMixin, models.Model):
